--- cypress.json
@@ -1,5 +1,0 @@
-{
-  "baseUrl": "http://localhost:8080",
-  "video": false,
-  "projectId": "jcb2wm"
-}
