+++ cypress/e2e/rest-variations.spec.js
